Add files via upload
This Stored Procedure Detects all the process from sysprocesses in SQL and captures the blocker process based on given condition. My condition is, if any process is blocking other process for more than 20 seconds then kill it and send the email to the user who created that process. 
To capture the Users information, i am using [Active Session] table in NAV.